finished histogram function.
todo:
Given that the user views a product's profile, they should be able to see the average rating for each predefined and custom parameter, clearly displayed alongside the other product details.
12: edit product
13:Given that the user submits a comment, the system should save the comment and display it on the product profile page, along with the user’s name and timestamp.
Given that multiple comments are submitted for a product, the system should display them in chronological order on the product profile page.