some minor fixes for acs
Given that the user logs out, when they revisit the application, then they should be prompted to log in again as the token has been cleared.
Given that a user attempts to upload an image, when the image type is not supported (e.g., non-JPG/PNG), then the system should display an appropriate error message.
Given the product listing page, when the user chooses "initial post time," then products should be sorted chronologically by their creation date.
Given that the user selects a specific category, when there are no products in that category, then the system should display an error message.
Given that a user is on the product ranking page when they are not selecting any particular category from a dropdown or filter, then they should see a list of top-rated products in all categories.